Move InternPool to PooledObjects namespace
